-Пофиксил баги с мертвой водой(добавил перегрузку MainCast в Artefact) -Исправил првиедение типов (uint) в LightningStaff -Исправил PoisonousSaliva, теперь не нужно вводить силу. -Убрал ненужные this в живой воде

+++ LightningStaff.cs
@@ -13,7 +13,7 @@ namespace GameByCash
         public override bool MainCast(Hero targetHero, uint strength)
         {
             if(ArtifactPower != 0
-            && ArtifactPower - strength >= 0//Может ли персонаж умереть от посоха?
+            && (int)ArtifactPower - (int)strength >= 0//Может ли персонаж умереть от посоха?
             && (int)targetHero.CurHlth - (int)strength >= 0)//если нет тогда нужна эта проверка иначе переделать чуть
             {
                 targetHero.CurHlth -= strength;
